Add files via upload
Preparation template is created to help set up the SAS data file for the individual text downloads for each individual year. It is recommended that these be run to create the SAS data sets prior to running all other code. Make sure to integrate the code provided by NHAMCS into this template to build each year.